chore: change pixiPointsRef to pixiPoints, reorganize imports
