+++ v3/src/components/graph/components/graph.tsx
@@ -15,6 +15,7 @@ import {DroppablePlot} from "./droppable-plot"
 import {ScatterDots} from "./scatterdots"
 import {CaseDots} from "./casedots"
 import {ChartDots} from "./chartdots"
+import { DotPlotDots } from "./dotplotdots"
 import {Marquee} from "./marquee"
 import {useDataSetContext} from "../../../hooks/use-data-set-context"
 import {isSetAttributeIDAction} from "../../data-display/models/display-model-actions"
@@ -38,7 +39,6 @@ import {mstReaction} from "../../../utilities/mst-reaction"
 import {onAnyAction} from "../../../utilities/mst-utils"
 import {IPixiPointsRef} from "../utilities/pixi-points"
 import {Adornments} from "../adornments/adornments"
-import { DotPlotDotsBaseComponent } from "./dotplotdots-base-component"
 
 import "./graph.scss"
 
@@ -181,8 +181,8 @@ export const Graph = observer(function Graph({graphController, graphRef, pixiPoi
       typeToPlotComponentMap = {
         casePlot: <CaseDots {...props}/>,
         dotChart: <ChartDots {...props}/>,
-        dotPlot: <DotPlotDotsBaseComponent {...props}/>,
-        scatterPlot: <ScatterDots {...props}/>,
+        dotPlot: <DotPlotDots {...props}/>,
+        scatterPlot: <ScatterDots {...props}/>
       }
     return typeToPlotComponentMap[plotType]
   }
