+++ v3/src/components/graph/models/graph-content-model.ts
@@ -19,10 +19,11 @@ import {defaultBackgroundColor} from "../../../utilities/color-utils"
 import {IGraphDataConfigurationModel} from "./graph-data-configuration-model"
 import {DataDisplayContentModel} from "../../data-display/models/data-display-content-model"
 import {GraphPlace} from "../../axis-graph-shared"
-import {axisPlaceToAttrRole, GraphAttrRole} from "../../data-display/data-display-types"
+import {axisPlaceToAttrRole, GraphAttrRole, PointDisplayType,
+        PointDisplayTypes} from "../../data-display/data-display-types"
 import {AxisPlace, AxisPlaces, ScaleNumericBaseType} from "../../axis/axis-types"
 import {kGraphTileType} from "../graph-defs"
-import {PlotType, PlotTypes, PointDisplayType, PointDisplayTypes} from "../graphing-types"
+import {PlotType, PlotTypes} from "../graphing-types"
 import {setNiceDomain} from "../utilities/graph-utils"
 import {GraphPointLayerModel, IGraphPointLayerModel, kGraphPointLayerType} from "./graph-point-layer-model"
 import {IAdornmentModel, IUpdateCategoriesOptions} from "../adornments/adornment-models"
@@ -116,6 +117,16 @@ export const GraphContentModel = DataDisplayContentModel
                                   dataset: IDataSet | undefined,
                                   attributeID: string | undefined): boolean {
       return self.dataConfiguration.placeCanAcceptAttributeIDDrop(place, dataset, attributeID)
+    },
+    nonDraggableAxisTicks(): { tickValues: number[], tickLabels: string[] } {
+      const tickValues: number[] = []
+      const tickLabels: string[] = []
+      const { binWidth, totalNumberOfBins  } = self.dataConfiguration.binDetails()
+      for (let i = 0; i < totalNumberOfBins; i++) {
+        tickValues.push(((i + 0.5) * binWidth))
+        tickLabels.push(`[${i * binWidth}, ${(i + 1) * binWidth})`)
+      }
+      return { tickValues, tickLabels }
     }
   }))
   .views(self => ({
