+++ v3/src/components/graph/utilities/dot-plot-utils.test.ts
@@ -3,9 +3,8 @@ import { calculatePointStacking, determineBinForCase } from "./dot-plot-utils"
 describe("determineBinForCase", () => {
   it("returns the bin number for a given case value in a binned dot plot", () => {
     const caseValue = 25
-    const totalNumberOfBins = 5
     const binWidth = 10
-    const binNumber = determineBinForCase(caseValue, totalNumberOfBins, binWidth)
+    const binNumber = determineBinForCase(caseValue, binWidth)
     expect(binNumber).toEqual(3)
   })
 })
