+++ v3/src/components/graph/utilities/pixi-points.ts
@@ -612,18 +612,18 @@ export class PixiPoints {
     })
   }
 
-  matchPointsToData(caseData: CaseData[], displayType: string, style: IPixiPointStyle) {
+  matchPointsToData(caseData: CaseData[], _displayType: string, style: IPixiPointStyle) {
     // If the display type has changed, we need to prepare for the transition between types
     // For now, the only display type values PixiPoints supports are "points" and "bars", so
     // all other display type values passed to this method will be treated as "points".
-    const useDisplayType = displayType !== "bars" ? "points" : "bars"
-    if (this.displayType !== useDisplayType) {
+    const displayType = _displayType !== "bars" ? "points" : "bars"
+    if (this.displayType !== displayType) {
       this.displayTypeTransitionState.isActive = true
       this.forEachPoint(point => {
         this.pointTransitionStates.set(point, { hasTransitioned: false })
       })
     }
-    this.displayType = useDisplayType
+    this.displayType = displayType
 
     // Stop here if we need to transition between display types. The need for a display type transition is only
     // triggered when the user changes the display type from points to bars or vice versa. Calls to `matchPointsToData`
